Rename SslByteHandler to StreamByteHandler.
The class can work with any .NET Stream and isn't limited to SslStream.
